refactoring of demo
